pyproject.toml
[project]
name = "rctp-pax-fcst"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "openpyxl>=3.1.5",
    "panda>=0.3.1",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "xlrd>=2.0.2",
]
